+++ sd_material_ui/Dialog.py
@@ -10,13 +10,8 @@ Material UI Dialog component
 Keyword arguments:
 - children (a list of or a singular dash component, string or number; optional): Children to render inside of the Dialog
 - id (string; required): Dialog ID
-- actions (optional): Actions component or list of components for the Dialog
+- ariaLabelledBy (string; default ''): List of space separated id's of elements to use as aria labels
 - className (string; default ''): CSS class name of the root element
-- modal (boolean; default False): Is the Dialog a modal (must click on an action to close)?
-
-NOTE: If the Dialog is created with modal=False, the callback that controls opening and closing
-the Dialog should include the component's open value as a state, and check whether it is either
-already open or closed before attempting to open or close it. See comments for the open prop.
 - open (boolean; default False): Is the dialog open?
 
 IMPORTANT: When using this component in Dash, a listener must be set up (either as state or
@@ -24,19 +19,18 @@ an input) for this component's props.open value in order to achieve the desired 
 If such a listener is not in place, the non-modal version of this dialog will contaminate
 other callbacks in the browser
 - autoScrollBodyContent (boolean; default False): If set to true, the body content of the Dialog will be scrollable.
-- actionsContainerClassName (string; optional): The className to add to the actions container's root element.
-- bodyClassName (string; optional): The className to add to the content's root element under the title.
-- contentClassName (string; optional): The className to add to the content container
-- overlayClassName (string; optional): The className to add to the Overlay component rendered behind the Dialog
-- paperClassName (string; optional): CSS class name of the Paper element
-- titleClassName (string; optional): The className to add to the title's root container element"""
+- componentContainerClassName (string; optional): The className to add to the component container
+- fullWidth (boolean; default True): The className to add to the content container
+- useBrowserSideClose (boolean; default False): If set to true, the Close Icon will show in the upper right corner of the dialog, closing the Dialog browser side
+- scroll (dict; default 'body'): "paper" or "body", Determines scroll container
+- style (dict; optional): Styles to be implemented as inline css"""
     @_explicitize_args
-    def __init__(self, children=None, id=Component.REQUIRED, actions=Component.UNDEFINED, className=Component.UNDEFINED, modal=Component.UNDEFINED, open=Component.UNDEFINED, autoScrollBodyContent=Component.UNDEFINED, actionsContainerClassName=Component.UNDEFINED, bodyClassName=Component.UNDEFINED, contentClassName=Component.UNDEFINED, overlayClassName=Component.UNDEFINED, paperClassName=Component.UNDEFINED, titleClassName=Component.UNDEFINED, **kwargs):
-        self._prop_names = ['children', 'id', 'actions', 'className', 'modal', 'open', 'autoScrollBodyContent', 'actionsContainerClassName', 'bodyClassName', 'contentClassName', 'overlayClassName', 'paperClassName', 'titleClassName']
+    def __init__(self, children=None, id=Component.REQUIRED, ariaLabelledBy=Component.UNDEFINED, className=Component.UNDEFINED, open=Component.UNDEFINED, autoScrollBodyContent=Component.UNDEFINED, componentContainerClassName=Component.UNDEFINED, fullWidth=Component.UNDEFINED, useBrowserSideClose=Component.UNDEFINED, scroll=Component.UNDEFINED, style=Component.UNDEFINED, actions=Component.UNDEFINED, **kwargs):
+        self._prop_names = ['children', 'id', 'ariaLabelledBy', 'className', 'open', 'autoScrollBodyContent', 'componentContainerClassName', 'fullWidth', 'useBrowserSideClose', 'scroll', 'style']
         self._type = 'Dialog'
         self._namespace = 'sd_material_ui'
         self._valid_wildcard_attributes =            []
-        self.available_properties = ['children', 'id', 'actions', 'className', 'modal', 'open', 'autoScrollBodyContent', 'actionsContainerClassName', 'bodyClassName', 'contentClassName', 'overlayClassName', 'paperClassName', 'titleClassName']
+        self.available_properties = ['children', 'id', 'ariaLabelledBy', 'className', 'open', 'autoScrollBodyContent', 'componentContainerClassName', 'fullWidth', 'useBrowserSideClose', 'scroll', 'style']
         self.available_wildcard_properties =            []
 
         _explicit_args = kwargs.pop('_explicit_args')