+++ sd_material_ui/_imports_.py
@@ -1,8 +1,10 @@
+from .Dialog import Dialog
 from .Drawer import Drawer
 from .RadioButtonGroup import RadioButtonGroup
 from .Toggle import Toggle
 
 __all__ = [
+    "Dialog",
     "Drawer",
     "RadioButtonGroup",
     "Toggle"