+++ usage.py
@@ -56,6 +56,19 @@ app.layout = html.Div([
         spacer,
 
         html.Div([
+            html.P([html.Strong('Test for dialog')]),
+            sd_material_ui.Dialog([
+                html.H3('Sample Dialog'),
+                html.Div(html.Button('Close Dialog'), id='closer')
+            ], id='output2'),
+            html.Div(id='input2', children=[
+                html.Button(children='Open the dialog')
+            ]),
+        ]),
+
+        spacer,
+
+        html.Div([
             html.P([html.Strong('Test for toggle switch')]),
             sd_material_ui.Toggle(
                 id='toggle-input',
@@ -64,7 +77,6 @@ app.layout = html.Div([
             ),
             html.P(id='toggle-output', children=['Flame off']),
         ]),
-
     ], style=dict(display='flex', flexWrap='wrap')),
 
     final_spacer,
@@ -130,21 +142,21 @@ app.layout = html.Div([
 #     return 'You have entered {}'.format(value)
 #
 #
-# # Callback for SDDialog (modal)
-# @app.callback(
-#     dash.dependencies.Output('output2', 'open'),
-#     [dash.dependencies.Input('input2', 'n_clicks'),
-#      dash.dependencies.Input('closer', 'n_clicks')],
-#     [dash.dependencies.State('output2', 'open')])
-# def show_modal_dialog(modal_click: int, close_button: int, open_state: bool):
-#     if modal_click and modal_click > 0:
-#         if not open_state:
-#             return True
-#     elif close_button:
-#         if open_state:
-#             return False
-#     else:
-#         return False
+# Callback for SDDialog (modal)
+@app.callback(
+    dash.dependencies.Output('output2', 'open'),
+    [dash.dependencies.Input('input2', 'n_clicks'),
+     dash.dependencies.Input('closer', 'n_clicks')],
+    [dash.dependencies.State('output2', 'open')])
+def show_modal_dialog(modal_click: int, close_button: int, open_state: bool):
+    if modal_click and modal_click > 0:
+        if not open_state:
+            return True
+    elif close_button:
+        if open_state:
+            return False
+    else:
+        return False
 #
 #
 # # Callback for SDDialog (non-modal)