+++ CHANGELOG.md
@@ -4,11 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 4.0.0 2020-08-11
 ### Changes
-- Changes the Drawer component to use the latest `material-ui`
 - Updates many dependencies to resolve various vulnerabilities
+- Changes the following components to use the latest `material-ui`
+    - Drawer
+    - Dialog
+    - RadioButtonGroup
 
 ### Related issues:
+- [402](https://github.com/StratoDem/sd-material-ui/issues/402)
 - [404](https://github.com/StratoDem/sd-material-ui/issues/404)
+- [412](https://github.com/StratoDem/sd-material-ui/issues/412)
 
 ## 3.2.0 2019-11-04
 ### Adds